Add files via upload
 In the current scenario,Online transactions have ended up a commonplace fashion now-a-days and security related to the identical is turning into a difficulty.
Face is an included internal organ whose random texture is complex, particular, and very stable at some stage in lifestyles, it can serve as a kind of living passport or password that one no longer to be remembered however can usually be a gift. So face popularity is one efficient manner of securing online transactions.
This makes security an important factor at the time of making digital payments. Hence, we propose a system to enhance the security of online transactions by providing a 2-step verification process: OTP verification followed by Facial Recognition.